+++ solarwinds_apm/trace/__init__.py
@@ -1,9 +1,7 @@
-from .inbound_metrics_processor import SolarWindsInboundMetricsSpanProcessor
 from .response_time_processor import ResponseTimeProcessor
 from .serviceentry_processor import ServiceEntrySpanProcessor
 
 __all__ = [
     "ServiceEntrySpanProcessor",
-    "SolarWindsInboundMetricsSpanProcessor",
     "ResponseTimeProcessor",
 ]
