+++ tests/unit/test_configurator/test_configurator_configure_otel.py
@@ -14,7 +14,6 @@ class TestConfiguratorConfigureOtelComponents:
         mock_apmconfig_enabled,
 
         mock_config_serviceentryid_processor,
-        mock_config_inbound_processor,
         mock_response_time_processor,
         mock_config_traces_exp,
         mock_config_metrics_exp,
@@ -29,9 +28,6 @@ class TestConfiguratorConfigureOtelComponents:
         )
 
         # mock_config_serviceentryid_processor.assert_called_once()
-        # mock_config_inbound_processor.assert_called_once_with(
-        #     mock_apmconfig_enabled,
-        # )
         # mock_response_time_processor.assert_called_once_with(
         #     mock_apmconfig_enabled,
         # )
@@ -52,7 +48,6 @@ class TestConfiguratorConfigureOtelComponents:
         mock_apmconfig_disabled,
 
         mock_config_serviceentryid_processor,
-        mock_config_inbound_processor,
         mock_response_time_processor,
         mock_config_traces_exp,
         mock_config_metrics_exp,
@@ -68,7 +63,6 @@ class TestConfiguratorConfigureOtelComponents:
         )
 
         mock_config_serviceentryid_processor.assert_not_called()
-        mock_config_inbound_processor.assert_not_called()
         mock_response_time_processor.assert_not_called()
         mock_config_traces_exp.assert_not_called()
         mock_config_metrics_exp.assert_not_called()
