+++ tests/unit/test_configurator/test_configurator_span_processors.py
@@ -27,95 +27,6 @@ class TestConfiguratorSpanProcessors:
         trace_mocks.get_tracer_provider.assert_called_once()
         trace_mocks.get_tracer_provider().add_span_processor.assert_called_once()
 
-    def test_configure_inbound_metrics_span_processor_no_exporters(
-        self,
-        mocker,
-    ):
-        # Save any exporters in os for later
-        old_exporter = os.environ.get("OTEL_TRACES_EXPORTER", None)
-        if old_exporter:
-            del os.environ["OTEL_TRACES_EXPORTER"]
-
-        trace_mocks = get_trace_mocks(mocker)
-        mock_processor = mocker.patch(
-            "solarwinds_apm.configurator.SolarWindsInboundMetricsSpanProcessor"
-        )
-
-        test_configurator = configurator.SolarWindsConfigurator()
-        test_configurator._configure_inbound_metrics_span_processor(
-            mocker.Mock(),
-        )  
-
-        trace_mocks.get_tracer_provider.assert_not_called()
-        trace_mocks.get_tracer_provider().add_span_processor.assert_not_called()
-        mock_processor.assert_not_called()
-
-        # Restore the os exporters
-        if old_exporter:
-            os.environ["OTEL_TRACES_EXPORTER"] = old_exporter
-
-    def test_configure_inbound_metrics_span_processor_no_sw_exporter(
-        self,
-        mocker,
-    ):
-        # Save any exporters in os for later
-        old_exporter = os.environ.get("OTEL_TRACES_EXPORTER", None)
-        if old_exporter:
-            del os.environ["OTEL_TRACES_EXPORTER"]
-        mocker.patch.dict(os.environ, {
-            "OTEL_TRACES_EXPORTER": "foo_exporter_not_sw",
-        })
-
-        trace_mocks = get_trace_mocks(mocker)
-        mock_processor = mocker.patch(
-            "solarwinds_apm.configurator.SolarWindsInboundMetricsSpanProcessor"
-        )
-
-        test_configurator = configurator.SolarWindsConfigurator()
-        test_configurator._configure_inbound_metrics_span_processor(
-            mocker.Mock(),
-        )  
-
-        trace_mocks.get_tracer_provider.assert_not_called()
-        trace_mocks.get_tracer_provider().add_span_processor.assert_not_called()
-        mock_processor.assert_not_called()
-
-        # Restore the os exporters
-        if old_exporter:
-            os.environ["OTEL_TRACES_EXPORTER"] = old_exporter
-
-    def test_configure_inbound_metrics_span_processor_ok(
-        self,
-        mocker,
-        mock_apmconfig_enabled,
-    ):
-        # Save any exporters in os for later
-        old_exporter = os.environ.get("OTEL_TRACES_EXPORTER", None)
-        if old_exporter:
-            del os.environ["OTEL_TRACES_EXPORTER"]
-        mocker.patch.dict(os.environ, {
-            "OTEL_TRACES_EXPORTER": "solarwinds_exporter",
-        })
-
-        trace_mocks = get_trace_mocks(mocker)
-        mock_processor = mocker.patch(
-            "solarwinds_apm.configurator.SolarWindsInboundMetricsSpanProcessor"
-        )
-
-        test_configurator = configurator.SolarWindsConfigurator()
-        test_configurator._configure_inbound_metrics_span_processor(
-            mock_apmconfig_enabled,
-        )       
-        trace_mocks.get_tracer_provider.assert_called_once()
-        trace_mocks.get_tracer_provider().add_span_processor.assert_called_once()
-        mock_processor.assert_called_once_with(
-            mock_apmconfig_enabled,
-        )
-
-        # Restore the os exporters
-        if old_exporter:
-            os.environ["OTEL_TRACES_EXPORTER"] = old_exporter
-
     def test_configure_response_time_processor_exporters_not_set(
         self,
         mocker,
